iterator: rename (Typed)RowIterator to Legacy(Typed)RowIterator
Like in the case of LegacyQueryResult, TypedRowIterator will be replaced
with a new one, and the old one preserved for compatibility reasons. As
a first step, it is renamed to LegacyTypedRowIterator.

Similarly, RowIterator is renamed to LegacyRowIterator to mark that it
is the part of the old API. There won't be a new RowIterator, though.