session: add methods for creating compatible session with different API
In order to make migration from the old API easier and allow doing it
gradually, some components of the client programs would probably like to
use the old API while the new components will use the new API. However,
in the current design, Session and LegacySession are two different types
and it's not possible to "cast" one to another - even though they have
nearly the same fields and implementations.

The previous commit made Cluster cloneable, based on the observation
that it's perfectly fine to clone Cluster's fields, construct a new one
and treat it as a shared facade, handle to the same "semantic" cluster.
The same applies to Session, actually - cloning a session would have
similar effect (though we encourage users to keep Session in an Arc so
that cloning is cheaper).

Instead of making GenericSession cloneable, we introduce methods which,
in reality, perform a clone but change the kind of session's API. This
allows to have two session objects which share the same resources but
have different APIs. This should be very useful when migrating large
projects to the new API - components that need to use the new API can
just "convert" the session to the new interface and use that.